// src-31052018-currently-working/app/plan/enums/plan-status.ts
/**
 * An enumeration representing the different plan statuses
 */
export enum PlanStatus {
  ACTIVE = 'Active',
  CANCELLED = 'Cancelled',
  IN_PROGRESS = 'In-Progress',
  IN_REVISION = 'In Revision',
  NEW = 'New',
  WILL_TERMINATE = 'Will Terminate',
  TERMINATED = 'Terminated',
  DELETED = 'Deleted'
}